+++ tango/local_workspace.py
@@ -504,6 +504,8 @@ class LocalWorkspace(Workspace):
 
     def registered_run(self, name: str) -> Dict[str, StepInfo]:
         run_dir = self.runs_dir / name
+        if not run_dir.is_dir():
+            raise KeyError(name)
         with SqliteDict(self.step_info_file, flag="r") as d:
             steps_for_run = {}
             for step_symlink in run_dir.iterdir():